+++ src/components/chat/chatList/ChatBody.tsx
@@ -1,10 +1,18 @@
+import { list } from '@/components/chat/ChatList';
 import ChatBox from '@/components/chat/chatList/ChatBox';
 import * as S from '../Chat.styles';
 
-const ChatBody = () => {
+type ChatBodyProps = {
+  chatList: list[];
+};
+
+const ChatBody = ({ chatList }: ChatBodyProps) => {
+  console.log('chatList1', chatList);
   return (
     <S.ChatBody>
-      <ChatBox />
+      {chatList?.map((item, idx) => {
+        return <ChatBox key={idx} lastChat={item.lastChat.content} sender={item.lastChat.sender} />;
+      })}
     </S.ChatBody>
   );
 };
