+++ src/components/chat/chatList/ChatBox.tsx
@@ -1,7 +1,28 @@
+import { useEffect, useState } from 'react';
+
 import Icon from '@/components/common/Icon';
 import * as S from '../Chat.styles';
 
-const ChatBox = () => {
+type ChatBoxProps = {
+  shopName: string;
+  lastChat: string;
+  sender: string;
+};
+
+const ChatBox = ({ shopName, lastChat, sender }: ChatBoxProps) => {
+  const [textCut, setTextCut] = useState<string>('');
+
+  /** textLengthCut20() : 마지막 채팅 미리보기 20글자까지만 보이게 */
+  const textLengthCut20 = (text: string) => {
+    const test = text.slice(0, 20);
+    if (test.length >= 15) setTextCut(test + ' ...');
+    else setTextCut(test);
+  };
+
+  useEffect(() => {
+    textLengthCut20(lastChat);
+  }, [lastChat]);
+
   return (
     <S.ChatBox>
       <div className="box_img_wrapper">
@@ -10,15 +31,16 @@ const ChatBox = () => {
       <div className="box_title_container">
         <div className="container">
           <div className="box_title_wrapper">
-            <span>어쩌구 쇼핑</span>
+            <span>0000 상품 문의</span>
           </div>
           <div className="box_subtitle_wrapper">
             <Icon name={'IconRate'} fill={'green'} width="10px" height="10px" />
             <span>12분 전</span>
+            <span>| {sender}</span>
           </div>
         </div>
         <div className="box_text_wrapper">
-          <span>어떤 문의이실까요?</span>
+          <span>{textCut}</span>
         </div>
       </div>
     </S.ChatBox>
